Added icons and log in view.

+++ Agent's App/Agent's App/Core/Repository/IRepository.cs
@@ -9,7 +9,7 @@ namespace Agent_s_App.Core.Repository
 {
 	public interface IRepository<TEntity> where TEntity : class
 	{
-		TEntity Get(int id);
+		TEntity Get(long id);
 		IEnumerable<TEntity> GetAll();
 		IEnumerable<TEntity> Find(Expression<Func<TEntity, bool>> predicate);
